Remove pyproject.toml (and poetry.lock)
They were causing a bug (or, at least, unexpected behavior) when
installing from pip:

When installing the library via pip, only the objective_turk package
would be installed, and the mturk package would be lost.

Removing pyproject.toml seems to fix this issue.

This project can probably be built with pyproject.toml and/or poetry,
but for the time being I'll stick with setup.py